+++ buffalogs/impossible_travel/alerting/alert_factory.py
@@ -6,10 +6,13 @@ from impossible_travel.alerting.base_alerting import BaseAlerting
 from impossible_travel.alerting.discord_alerting import DiscordAlerting
 from impossible_travel.alerting.dummy_alerting import DummyAlerting
 from impossible_travel.alerting.email_alerting import EmailAlerting
+from impossible_travel.alerting.googlechat_alerting import GoogleChatAlerting
 from impossible_travel.alerting.http_request import HTTPRequestAlerting
-from impossible_travel.alerting.microsoft_teams_alerter import MicrosoftTeamsAlerting
+from impossible_travel.alerting.mattermost_alerting import MattermostAlerting
+from impossible_travel.alerting.microsoft_teams_alerting import MicrosoftTeamsAlerting
 from impossible_travel.alerting.pushover_alerting import PushoverAlerting
-from impossible_travel.alerting.slack_alerter import SlackAlerter
+from impossible_travel.alerting.rocketchat_alerting import RocketChatAlerting
+from impossible_travel.alerting.slack_alerting import SlackAlerting
 from impossible_travel.alerting.telegram_alerting import TelegramAlerting
 from impossible_travel.alerting.webhook import WebHookAlerting
 
@@ -45,7 +48,7 @@ class AlertFactory:
             case BaseAlerting.SupportedAlerters.DUMMY:
                 return DummyAlerting(self.alert_config)
             case BaseAlerting.SupportedAlerters.SLACK:
-                return SlackAlerter(self.alert_config)
+                return SlackAlerting(self.alert_config)
             case BaseAlerting.SupportedAlerters.WEBHOOK:
                 return WebHookAlerting(self.alert_config)
             case BaseAlerting.SupportedAlerters.HTTPREQUEST:
@@ -60,5 +63,11 @@ class AlertFactory:
                 return DiscordAlerting(self.alert_config)
             case BaseAlerting.SupportedAlerters.MICROSOFTTEAMS:
                 return MicrosoftTeamsAlerting(self.alert_config)
+            case BaseAlerting.SupportedAlerters.ROCKETCHAT:
+                return RocketChatAlerting(self.alert_config)
+            case BaseAlerting.SupportedAlerters.GOOGLECHAT:
+                return GoogleChatAlerting(self.alert_config)
+            case BaseAlerting.SupportedAlerters.MATTERMOST:
+                return MattermostAlerting(self.alert_config)
             case _:
                 raise ValueError(f"Unsupported alerter: {self.active_alerter}")