+++ buffalogs/impossible_travel/alerting/base_alerting.py
@@ -1,6 +1,10 @@
+import json
 import logging
+import os
 from abc import ABC, abstractmethod
 from enum import Enum
+
+from django.conf import settings
 
 
 class BaseAlerting(ABC):
@@ -18,6 +22,9 @@ class BaseAlerting(ABC):
         PUSHOVER = "pushover"
         DISCORD = "discord"
         MICROSOFTTEAMS = "microsoftteams"
+        ROCKETCHAT = "rocketchat"
+        GOOGLECHAT = "googlechat"
+        MATTERMOST = "mattermost"
 
     def __init__(self):
         super().__init__()
@@ -30,3 +37,13 @@ class BaseAlerting(ABC):
         Must be implemented by concrete classes.
         """
         raise NotImplementedError
+
+    @staticmethod
+    def read_config(alerter_key: str):
+        """
+        Read the configuration for a specific alerter from alerting.json.
+        """
+        config_path = os.path.join(settings.CERTEGO_BUFFALOGS_CONFIG_PATH, "buffalogs/alerting.json")
+        with open(config_path, mode="r", encoding="utf-8") as f:
+            config = json.load(f)
+        return config.get(alerter_key, {})