+++ buffalogs/impossible_travel/alerting/discord_alerting.py
@@ -7,7 +7,7 @@ from impossible_travel.models import Alert
 
 class DiscordAlerting(BaseAlerting):
     """
-    Concrete implementation of the BaseQuery class for Discord.
+    Concrete implementation of the BaseQuery class for DiscordAlerting.
     """
 
     def __init__(self, alert_config: dict):
@@ -15,8 +15,12 @@ class DiscordAlerting(BaseAlerting):
         Constructor for the Discord Alerter query object.
         """
         super().__init__()
-        # here we can access the alert_config to get the configuration for the alerter
         self.webhook_url = alert_config.get("webhook_url")
+        self.username = alert_config.get("username")
+
+        if not self.webhook_url or not self.username:
+            self.logger.error("Discord Alerter configuration is missing required fields.")
+            raise ValueError("Discord Alerter configuration is missing required fields.")
 
     def notify_alerts(self):
         """
@@ -24,18 +28,17 @@ class DiscordAlerting(BaseAlerting):
         """
         alerts = Alert.objects.filter(notified=False)
         for alert in alerts:
-            # the alert message to send
             alert_msg = f"Dear user,\n\nAn unusual login activity has been detected:\n\n{alert.description}\n\nStay Safe,\nBuffalogs"
-
             discord_message = {
-                "username": "BuffaLogs Alert",
+                "username": self.username,
                 "embeds": [{"title": f"Login Anomaly Alert: {alert.name}", "description": alert_msg, "color": 16711680}],  # red
             }
-
             headers = {"Content-Type": "application/json"}
-
-            requests.post(self.webhook_url, headers=headers, data=json.dumps(discord_message))
-
-            self.logger.info("Alerting %s", alert.name)
-            alert.notified = True
-            alert.save()
+            try:
+                resp = requests.post(self.webhook_url, headers=headers, data=json.dumps(discord_message))
+                resp.raise_for_status()
+                self.logger.info(f"Discord alert sent: {alert.name}")
+                alert.notified = True
+                alert.save()
+            except requests.RequestException as e:
+                self.logger.exception(f"Discord alert failed for {alert.name}: {str(e)}")