+++ buffalogs/impossible_travel/alerting/pushover_alerting.py
@@ -5,7 +5,7 @@ from impossible_travel.models import Alert
 
 class PushoverAlerting(BaseAlerting):
     """
-    Concrete implementation of the BaseQuery class for Pushover.
+    Concrete implementation of the BaseQuery class for PushoverAlerting.
     """
 
     def __init__(self, alert_config: dict):
@@ -13,9 +13,12 @@ class PushoverAlerting(BaseAlerting):
         Constructor for the Pushover Alerter query object.
         """
         super().__init__()
-        # here we can access the alert_config to get the configuration for the alerter
         self.api_key = alert_config.get("api_key")
         self.user_key = alert_config.get("user_key")
+
+        if not self.api_key or not self.user_key:
+            self.logger.error("Pushover Alerter configuration is missing required fields.")
+            raise ValueError("Pushover Alerter configuration is missing required fields.")
 
     def notify_alerts(self):
         """
@@ -23,17 +26,15 @@ class PushoverAlerting(BaseAlerting):
         """
         alerts = Alert.objects.filter(notified=False)
         for alert in alerts:
-            # the alert message to send
             alert_msg = (
                 f"Login Anomaly Alert: {alert.name}\nDear user,\n\nAn unusual login activity has been detected:\n\n{alert.description}\n\nStay Safe,\nBuffalogs"
             )
-
-            # create the payload to send
             payload = {"token": self.api_key, "user": self.user_key, "message": alert_msg}
-
-            # post the request to pushover
-            requests.post("https://api.pushover.net/1/messages.json", data=payload)
-
-            self.logger.info("Alerting %s", alert.name)
-            alert.notified = True
-            alert.save()
+            try:
+                resp = requests.post("https://api.pushover.net/1/messages.json", data=payload)
+                resp.raise_for_status()
+                self.logger.info(f"Pushover alert sent: {alert.name}")
+                alert.notified = True
+                alert.save()
+            except requests.RequestException as e:
+                self.logger.exception(f"Pushover alert failed for {alert.name}: {str(e)}")