+++ buffalogs/impossible_travel/tests/test_alert_discord.py
@@ -1,7 +1,9 @@
 import json
 from unittest.mock import MagicMock, patch
 
+import requests
 from django.test import TestCase
+from impossible_travel.alerting.base_alerting import BaseAlerting
 from impossible_travel.alerting.discord_alerting import DiscordAlerting
 from impossible_travel.models import Alert, Login, User
 
@@ -10,7 +12,7 @@ class TestDiscordAlerting(TestCase):
     def setUp(self):
         """Set up test data before running tests."""
 
-        self.discord_config = {"webhook_url": "https://discord.com/api/webhooks/WEBHOOK"}
+        self.discord_config = BaseAlerting.read_config("discord")
         self.discord_alerting = DiscordAlerting(self.discord_config)
 
         # Create a dummy user
@@ -30,7 +32,7 @@ class TestDiscordAlerting(TestCase):
         self.discord_alerting.notify_alerts()
 
         expected_payload = {
-            "username": "BuffaLogs Alert",
+            "username": self.discord_config["username"],
             "embeds": [
                 {
                     "title": "Login Anomaly Alert: Imp Travel",
@@ -41,11 +43,35 @@ class TestDiscordAlerting(TestCase):
         }
 
         mock_post.assert_called_once_with(
-            "https://discord.com/api/webhooks/WEBHOOK",
+            self.discord_config["webhook_url"],
             headers={"Content-Type": "application/json"},
             data=json.dumps(expected_payload),
         )
 
+    @patch("requests.post")
+    def test_no_alerts(self, mock_post):
+        """Test that no alerts are sent when there are no alerts to notify"""
+        Alert.objects.all().update(notified=True)
+        self.discord_alerting.notify_alerts()
+        self.assertEqual(mock_post.call_count, 0)
+
+    def test_improper_config(self):
+        """Test that an error is raised if the configuration is not correct"""
+        with self.assertRaises(ValueError):
+            DiscordAlerting({})
+
+    @patch("requests.post")
+    def test_alert_network_failure(self, mock_post):
+        """Test that alert is not marked as notified if there are any Network Fails"""
+        # Simulate network/API failure
+        mock_post.side_effect = requests.RequestException()
+
+        self.discord_alerting.notify_alerts()
+
+        # Reload the alert from DB to check its state
+        alert = Alert.objects.get(pk=self.alert.pk)
+        self.assertFalse(alert.notified)
+
     def test_send_actual_alert(self):
-        # Sending actual alert message to the chat
+        """Test sending an actual alert"""
         self.discord_alerting.notify_alerts()